Style: Adjust the Add Product hover and Logo area
I added a hover effect for the add product by the Sales Analytics
Section. It doesn't look nice without its hover effect.

I also removed the white background behind the logo area. I used it to
see the areas that section covers. It didn't look good to me so I took t
out.

Yeah. PS: I removed the comment section of the table in the HTML so that
the table is dynamically populated without a pre-defined template.